+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Fix #17. The `AlignMultilineParametersFixer` doesn't account for the & reference token in function parameters (firecow).
 
 ## [1.3.1] - 2025-11-24
 ### Fixed